Assignment-class-8/requirements.txt
streamlit
google-generativeai
python-dotenv
gtts
